@@ -17,12 +17,12 @@
     "typescript": "3.5.3"
   },
   "devDependencies": {
-    "@types/node": "12.6.8",
+    "@types/node": "12.7.2",
     "ts-loader": "6.0.4",
     "ts-node": "8.3.0",
-    "tslint": "5.18.0",
-    "webpack": "4.38.0",
-    "webpack-cli": "3.3.6",
+    "tslint": "5.19.0",
+    "webpack": "4.39.2",
+    "webpack-cli": "3.3.7",
     "webpack-node-externals": "1.7.2"
   }
 }
