@@ -1,6 +1,5 @@
 import { HttpException } from './http.exception';
 import { HttpStatus } from '../enums/http-status.enum';
-import { createHttpExceptionBody } from '../utils/http-exception-body.util';
 
 export class InternalServerErrorException extends HttpException {
   constructor(
@@ -8,7 +7,7 @@ export class InternalServerErrorException extends HttpException {
     error = 'Internal Server Error',
   ) {
     super(
-      createHttpExceptionBody(message, error, HttpStatus.INTERNAL_SERVER_ERROR),
+      HttpException.createBody(message, error, HttpStatus.INTERNAL_SERVER_ERROR),
       HttpStatus.INTERNAL_SERVER_ERROR,
     );
   }
