@@ -1,9 +1,9 @@
 import { RuntimeException } from './runtime.exception';
 
 export class UnknownElementException extends RuntimeException {
-  constructor() {
+  constructor(name?: string) {
     super(
-      'Nest could not find the given element (this class does not exist in the current context)',
+      `Nest could not find ${name || 'given'} element (this provider does not exist in the current context)`,
     );
   }
 }
