@@ -9,6 +9,7 @@ import {
   CUSTOM_ROUTE_AGRS_METADATA,
   HEADERS_METADATA,
   HTTP_CODE_METADATA,
+  REDIRECT_METADATA,
   RENDER_METADATA,
   ROUTE_ARGS_METADATA,
 } from '@nestjs/common/constants';
@@ -36,6 +37,7 @@ import { PipesContextCreator } from '../pipes/pipes-context-creator';
 import { IRouteParamsFactory } from './interfaces/route-params-factory.interface';
 import {
   CustomHeader,
+  RedirectResponse,
   RouterResponseController,
 } from './router-response-controller';
 
@@ -192,9 +194,11 @@ export class RouterExecutionContext {
         type === RouteParamtypes.RESPONSE || type === RouteParamtypes.NEXT,
     );
 
+    const httpRedirectResponse = this.reflectRedirect(callback);
     const fnHandleResponse = this.createHandleResponseFn(
       callback,
       isResponseHandled,
+      httpRedirectResponse,
     );
 
     const httpCode = this.reflectHttpStatusCode(callback);
@@ -204,7 +208,6 @@ export class RouterExecutionContext {
 
     const responseHeaders = this.reflectResponseHeaders(callback);
     const hasCustomHeaders = !isEmpty(responseHeaders);
-
     const handlerMetadata: HandlerMetadata = {
       argsLength,
       fnHandleResponse,
@@ -216,6 +219,10 @@ export class RouterExecutionContext {
     };
     this.handlerMetadataStorage.set(instance, methodName, handlerMetadata);
     return handlerMetadata;
+  }
+
+  public reflectRedirect(callback: (...args: any[]) => any): RedirectResponse {
+    return Reflect.getMetadata(REDIRECT_METADATA, callback);
   }
 
   public reflectHttpStatusCode(callback: (...args: any[]) => any): number {
@@ -358,13 +365,18 @@ export class RouterExecutionContext {
   public createHandleResponseFn(
     callback: (...args: any[]) => any,
     isResponseHandled: boolean,
+    redirectResponse?: RedirectResponse,
     httpStatusCode?: number,
   ) {
     const renderTemplate = this.reflectRenderTemplate(callback);
-
     if (renderTemplate) {
       return async <TResult, TResponse>(result: TResult, res: TResponse) => {
         await this.responseController.render(result, res, renderTemplate);
+      };
+    }
+    if (redirectResponse && redirectResponse.url) {
+      return async <TResult, TResponse>(result: TResult, res: TResponse) => {
+        await this.responseController.redirect(result, res, redirectResponse);
       };
     }
     return async <TResult, TResponse>(result: TResult, res: TResponse) => {
